requirements.txt
isort
pylint
Pillow